+++ changelog.md
@@ -3,6 +3,10 @@
 
 > **Note:** The dates included here are the dates of the commits that were made to the repository. The actual release dates may be different.
 
+# 27 April 2025
+* Include support for results and trace to include the navigation path to the raw resource content
+* Update the fhirpath.js library to 4.0.0 (and several other packages)
+
 # 21 January 2025
 * Add exception checking when trying to write to the localstorage (discovered while parsing a HUGE json bundle)